Only query the shader type if control is still valid
This fixes an out-of-scope access when shutting down in
fullscreen OpenGL mode.

Context: Previously, the global config record was
constructed within SDL main's `try { ... }` block, which
goes out-of-scope when an exception is thrown. Attempting
to access any configuration records after the exception
(such as when changing video modes to shutdown) would
result in accessing previously destroyed memory.