Check seek and tell return values in drive overlay (CWE-252), (CWE-687)
from Coverity:

The function returns a value that indicates an error condition.
If this is not checked, the error condition may not be handled
correctly.

In OverlayFile::​create_copy(): Value returned from a library
function is not checked for errors before being used. This
value may indicate an error condition. (CWE-252)

Check the return value from ftell in drive overlay (CWE-687)

and ...

The negative argument will be interpreted as a very large
unsigned value.

In OverlayFile::​create_copy(): Negative value used as argument
to a function expecting a positive value (for example, size of
buffer or allocation) (CWE-687)