Check the return value from seek in pcjr boot (CWE-252)
From Coverity:

The function returns a value that indicates an error condition.
If this is not checked, the error condition may not be handled
correctly.

In BOOT::​Run(): Value returned from a library function is not
checked for errors before being used. This value may indicate
an error condition. (CWE-252)