Run the GameBlaster's channel at the mixer's rate
Previously the SAA-1099 had an output sample rate of
~27 KHz based on its hardcoded clock/256 divisor.

Context:
  - The SAA-1099's are clocked at 1'4318'180 Hz / 2
  - And MAME's code divided that by 256 = 27'965 Hz

This commit makes that 256 clock divisor flexible by
taking it in as a variable in the device's constructor.

This commit then passes in 32 instead, giving us an
output rate of (1'4318'180 Hz / 2) / 32 = 223'721 Hz,
which is then sampled down to the mixer's native rate.