Import ripa's PC Speaker patch
Patch notes from ripa (2011-10-29, 07:49):

I implemented better interpolation/filtering. Now Hard
Drivin' 2 intro sounds pretty good. Perestroika intro also
sounds almost exactly the same as the real thing. The
improvement is mostly audible in games that use the PC
speaker for outputing digitized sound.

If you can hear a high-pitched whine in such games, you can
get rid of it by using a lower pcrate in dosbox.conf (16000
for Wizball for example, 32000 for StarCon2 and Perestroika
- you should also lower mixer rate to match).

There are some lingering issues. First of all, when
playing, you might hear slight crackling. However, if you
capture the sound with Ctrl+F6, you can't hear it in the
captured version. This leads me to believe that the
callback with the new filtering takes too long to execute
sometimes.

The second issue is that for speed reasons, I had to use a
lookup table for the filtering, instead of evaluating the
"impulse" function directly. This leads to worse sound
quality compared to the reference implementation that
doesn't use the lookup table. Also, there's some hidden bug
in the lookup implementation, because if I increase the
lookup table size to increase quality, it breaks
completely.

The final issue is that the mixer sometimes request more
samples that I expect it to. There's a hack for that, but
it results in a couple of "too many samples" prints when
you start up dosbox. The pc speaker code assumes that the
mixer callback is called every emulated millisecond. I
don't know how the mixer works, so I don't know if that
assumption is correct.

Imported-from: https://www.vogons.org/viewtopic.php?p=247154#p247154