Standardize the Game Blaster's rendering routines
Adds a 'ConfigureRelease' function to the soft-limiter to
properly scale the release amount if the frame-rate and
frame-per-processing event differs from the "typical" baseline
(which is taken from the Gravis Ultrasound's volume ramp).

Previously, it was assumed the limiter would be called on a
stream opearting at 44.1 KHz (or 48 KHz) *and* when given
roughly one millisecond worth of frames per processing event.

With the addition of higher-frequency sources operating and
processing at per-sample rates, the release amount was no
longer correct (and far too agressive).

Thanks to John Novak for reporting this.