Improve keyboard layout logic
The previous country-code lookup returned US on fail, which was
ambiguous if the user actually wanted the US.  This commit now
returns a bool.

The previous DOS_LoadKeyboardLayoutFromLanguage function only
checked if the country-code was valid after getting it from
either the setup or the layout.

This commit now checks the country code after getting it from
setup, and if that's invalid then tries from the layout.