Mark all core run functions with noexcept
The cores do not use, handle, or throw exceptions.

There were only a couple uses of the "new" operator,
which can throw if it fails to allocate memory. So this
commit move those to the no-throwing variant.

The core themselves only use return codes to manage signal
handling and control flow (both good and bad).  If things
go off the rails, they quit immediately with E_Exit.