Refactor BatchFile::ReadLine()
Replace char buffers with std::string.

Split up into a few functions.

Simplify the control flow.

Remove the cycle hack, as it doesn't appear to do much.
Cycles can be set within batch files without it.

Properly identify labels and comments when preceeded by tabs,
equals signs, and spaces.

Truncate commands that are too long instead of crashing.
Later commits should refactor the function and the code that calls
it so that it uses a std::string instead of a fixed size
char buffer.