Fix a potential TSAN race when closing FluidSynth
A TSAN data race is sometimes triggered on close involving access
to the static vector of data written into by the service.

This commit mirrors the MT-32's closing sequence where we:

 1. stop rendering and drain the channel
 2. shut down (and delete) the library's service
 3. shutdown (and delete) the audio channel

This sequence eliminates any possibility that the service might
still be pointing to or accessing the memory in which the channel
might also still be playing.

Test with TSAN
~~~~~~~~~~~~~~

Setup:

  meson setup -Dbuildtype=debug -Db_sanitize=thread \
  --force-fallback-for=fluidsynth,mt32emu build/tsan-midi

Compile:

  meson compile -C build/tsan-midi

Binary:

  build/tsan-midi/dosbox