Normalise Windows context menus created by the installer
From now on, right-clicking either on a folder or on the empty space
inside a folder in Windows Explorer will behave the same way: start
DOSBox with the working dir set to that folder.

The previous ability to execute .exe, .com and .bat files in DOSBox
using the right-click context menu has been removed as it's
inconsistent, confusing, and has limited applicability. It's just not
worth it for the confusion it introduces.

The ability to execute a .conf file using the right-click context menu
has been retained.

ping ci