Let the VGA-machine's double-scanned lines be duplicated or drawn
Context:

When in low resolution (320x200) EGA and VGA modes, the source content
is only 320x200 pixels that VGA cards' "double-scanned" up to 640x400,
producing two lines for every original line.

Because these are duplicates, DOSBox (and Staging) previously saved
processing and memory bandwidth (more FPS!) by letting the rendering
side scale the 320x200 scene up to the desired output side.

This works fine when the content is linearly scaled without concern
for deliniating individual lines.

However, when it comes to OpenGL shaders designed to emulate
individual phosphors and lines as drawn on a CRT, having all the
original lines as drawn by the VGA card becomes much more important.
We /want/ the shader applied to each of these original 400 lines to
accurately emulate the VGA's double-scanned signal hitting the CRT.

So that's what this commit adds:

 1) We can tell the VGA side if we don't care about lines (ie: duplicate)
    when we know the content is being linearly scaled

 2) We can ask for all 400 lines when we're confident shaders are
    being used (and the user wants to see these details).